+++ ObjectRenderers/ConsoleRendererBase.cs
@@ -1,4 +1,5 @@
 using System.Text;
+using BoxOfYellow.ConsoleMarkdownRenderer.Styling;
 using Markdig.Helpers;
 using Markdig.Renderers;
 using Markdig.Syntax;
@@ -152,7 +153,13 @@ namespace BoxOfYellow.ConsoleMarkdownRenderer.ObjectRenderers
         }
 
         protected void AddThematicBreakImplementation()
-            => m_frames.Peek().AddRow(new Rule());
+        {
+            var rule = string.IsNullOrEmpty(Options.ThematicBreakTitle)
+                ? new Rule()
+                : new Rule(Options.ThematicBreakTitle);
+            rule.Style = Options.ThematicBreak.ToSpectreStyle();
+            m_frames.Peek().AddRow(rule);
+        }
 
         /// <summary>
         /// Adds an arbitrary <see cref="IRenderable"/> as a row in the current frame.